requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
python-dotenv>=1.0.0
groq>=0.4.0
python-dateutil>=2.8.0
pydantic>=2.5.0
aiohttp>=3.9.0
openai>=1.12.0
tenacity>=8.2.0
structlog>=24.1.0

